removed gcc 4.1 compiler warnings on linux at daimi
Ignore-this: b161f54e8dbc4139e21787056fadc533

darcs-hash:20100803152027-3a0eb-dbdff2f61e925d19058fa9bd246d8c861f93a37b.gz

+++ Display/HUD.cpp
@@ -83,8 +83,8 @@ void HUD::Handle(RenderingEventArg arg) {
 
         x = (*itr)->x;
         y = (*itr)->y;
-        w = (*itr)->texr->GetWidth() * (*itr)->scaleX;
-        h = (*itr)->texr->GetHeight() * (*itr)->scaleY;
+        w = (unsigned int)((*itr)->texr->GetWidth() * (*itr)->scaleX);
+        h = (unsigned int)((*itr)->texr->GetHeight() * (*itr)->scaleY);
         if (w*h == 0) continue;
 
         n  = Vector<3,float>(0,0,1);
